ocr/requirements.txt
anyio==4.5.0
astor==0.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
exceptiongroup==1.2.2
fire==0.6.0
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
imageio==2.35.1
imgaug==0.4.0
importlib_resources==6.4.5
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
matplotlib==3.7.5
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.2
pillow==10.4.0
protobuf==5.28.2
pyclipper==1.3.0.post5
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyWavelets==1.4.1
PyYAML==6.0.2
rapidfuzz==3.9.7
requests==2.32.3
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.4.0
tifffile==2023.7.10
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
